+++ src/geom/objects/triangle.rs
@@ -13,13 +13,13 @@ pub struct Triangle {
 }
 
 impl Triangle {
+    #[deprecated(
+        since = "0.4.0-alpha0.5",
+        note = "Use another collision library like `vek` instead; please comment on issue #552 for use-cases other libraries don't solve"
+    )]
     ///Create a triangle from `Vector`s of all three points
-    pub fn new(a: impl Into<Vector>, b: impl Into<Vector>, c: impl Into<Vector>) -> Triangle {
-        Triangle {
-            a: a.into(),
-            b: b.into(),
-            c: c.into(),
-        }
+    pub fn new(a: Vector, b: Vector, c: Vector) -> Triangle {
+        Triangle { a, b, c }
     }
     ///Calculate the area of the triangle
     pub fn area(self) -> f32 {