+++ src/geom/shape.rs
@@ -1,10 +1,12 @@
+#![allow(deprecated)]
+
 use crate::geom::{about_equal, Circle, Line, Rectangle, Transform, Triangle, Vector};
 
 /// The collision and positional attributes of shapes
 pub trait Shape {
     /// If the point lies on the shape's boundary or within it
     #[must_use]
-    fn contains(&self, point: impl Into<Vector>) -> bool;
+    fn contains(&self, point: Vector) -> bool;
     /// If any area bounded by the shape falls on the line
     #[must_use]
     fn intersects(&self, line: &Line) -> bool {
@@ -22,6 +24,10 @@ pub trait Shape {
     }
     /// If any area is bounded by both either shape
     #[must_use]
+    #[deprecated(
+        since = "0.4.0-alpha0.5",
+        note = "Use another collision library like `vek` instead; please comment on issue #552 for use-cases other libraries don't solve"
+    )]
     fn overlaps(&self, other: &impl Shape) -> bool;
 
     /// The point all other points are equidistant to in the shape
@@ -34,6 +40,10 @@ pub trait Shape {
     ///
     /// Note: if you want to get the collision of rotated shapes, you probably
     /// want to use [ncollide](https://crates.io/crates/ncollide)
+    #[deprecated(
+        since = "0.4.0-alpha0.5",
+        note = "Use another collision library like `vek` instead; please comment on issue #552 for use-cases other libraries don't solve"
+    )]
     #[must_use]
     fn transformed_bounding_box(&self, transform: Transform) -> Rectangle {
         let bb = self.bounding_box();
@@ -43,8 +53,8 @@ pub trait Shape {
             * Transform::translate(-bb.pos - bb.center());
         // Get new corner position
         let tl = transform * bb.pos;
-        let tr = transform * (bb.pos + Vector::new(bb.size.x, 0));
-        let bl = transform * (bb.pos + Vector::new(0, bb.size.y));
+        let tr = transform * (bb.pos + Vector::new(bb.size.x, 0.0));
+        let bl = transform * (bb.pos + Vector::new(0.0, bb.size.y));
         let br = transform * (bb.pos + bb.size);
         // Get min and max points
         let min = tr.min(tl).min(bl).min(br);
@@ -54,7 +64,7 @@ pub trait Shape {
     }
     /// Create a copy of the shape with an offset center
     #[must_use]
-    fn translate(&self, amount: impl Into<Vector>) -> Self
+    fn translate(&self, amount: Vector) -> Self
     where
         Self: Sized;
     /// Create a copy of the shape that is contained within the bound
@@ -72,17 +82,17 @@ pub trait Shape {
     }
     /// Create a copy of the shape with an offset center
     #[must_use]
-    fn with_center(&self, center: impl Into<Vector>) -> Self
+    fn with_center(&self, center: Vector) -> Self
     where
         Self: Sized,
     {
-        self.translate(center.into() - self.center())
+        self.translate(center - self.center())
     }
 }
 
 impl Shape for Circle {
-    fn contains(&self, v: impl Into<Vector>) -> bool {
-        (v.into() - self.center()).len2() < self.radius.powi(2)
+    fn contains(&self, v: Vector) -> bool {
+        (v - self.center()).len2() < self.radius.powi(2)
     }
     fn overlaps_circle(&self, c: &Circle) -> bool {
         (self.center() - c.center()).len2() < (self.radius + c.radius).powi(2)
@@ -97,21 +107,19 @@ impl Shape for Circle {
     fn bounding_box(&self) -> Rectangle {
         Rectangle::new(
             self.pos - Vector::ONE * self.radius,
-            Vector::ONE * 2 * self.radius,
+            Vector::ONE * 2.0 * self.radius,
         )
     }
-    fn translate(&self, v: impl Into<Vector>) -> Self {
+    fn translate(&self, v: Vector) -> Self {
         Circle {
-            pos: self.pos + v.into(),
+            pos: self.pos + v,
             radius: self.radius,
         }
     }
 }
 
 impl Shape for Rectangle {
-    fn contains(&self, point: impl Into<Vector>) -> bool {
-        let p = point.into();
-
+    fn contains(&self, p: Vector) -> bool {
         p.x >= self.x()
             && p.y >= self.y()
             && p.x < self.x() + self.width()
@@ -139,22 +147,25 @@ impl Shape for Rectangle {
     }
 
     fn center(&self) -> Vector {
-        self.pos + self.size / 2
+        self.pos + self.size / 2.0
     }
     fn bounding_box(&self) -> Rectangle {
         *self
     }
-    fn translate(&self, v: impl Into<Vector>) -> Self {
+    fn translate(&self, v: Vector) -> Self {
         Rectangle {
-            pos: self.pos + v.into(),
+            pos: self.pos + v,
             size: self.size,
         }
     }
 }
 
+#[deprecated(
+    since = "0.4.0-alpha0.5",
+    note = "Use another collision library like `vek` instead; please comment on issue #552 for use-cases other libraries don't solve"
+)]
 impl Shape for Triangle {
-    fn contains(&self, v: impl Into<Vector>) -> bool {
-        let v = v.into();
+    fn contains(&self, v: Vector) -> bool {
         // form three triangles with this new vector
         let t_1 = Triangle::new(v, self.a, self.b);
         let t_2 = Triangle::new(v, self.b, self.c);
@@ -190,15 +201,14 @@ impl Shape for Triangle {
     }
 
     fn center(&self) -> Vector {
-        (self.a + self.b + self.c) / 3
+        (self.a + self.b + self.c) / 3.0
     }
     fn bounding_box(&self) -> Rectangle {
         let min = self.a.min(self.b.min(self.c));
         let max = self.a.max(self.b.max(self.c));
         Rectangle::new(min, max - min)
     }
-    fn translate(&self, v: impl Into<Vector>) -> Self {
-        let v = v.into();
+    fn translate(&self, v: Vector) -> Self {
         Triangle {
             a: self.a + v,
             b: self.b + v,
@@ -207,9 +217,12 @@ impl Shape for Triangle {
     }
 }
 
+#[deprecated(
+    since = "0.4.0-alpha0.5",
+    note = "Use another collision library like `vek` instead; please comment on issue #552 for use-cases other libraries don't solve"
+)]
 impl Shape for Line {
-    fn contains(&self, v: impl Into<Vector>) -> bool {
-        let v = v.into();
+    fn contains(&self, v: Vector) -> bool {
         about_equal(
             v.distance(self.a) + v.distance(self.b),
             self.a.distance(self.b),
@@ -266,15 +279,14 @@ impl Shape for Line {
     }
 
     fn center(&self) -> Vector {
-        (self.a + self.b) / 2
+        (self.a + self.b) / 2.0
     }
     fn bounding_box(&self) -> Rectangle {
         let min = self.a.min(self.b);
         let max = self.a.max(self.b);
         Rectangle::new(min, max - min)
     }
-    fn translate(&self, v: impl Into<Vector>) -> Self {
-        let v = v.into();
+    fn translate(&self, v: Vector) -> Self {
         Line {
             a: self.a + v,
             b: self.b + v,
@@ -284,8 +296,8 @@ impl Shape for Line {
 }
 
 impl Shape for Vector {
-    fn contains(&self, v: impl Into<Vector>) -> bool {
-        *self == v.into()
+    fn contains(&self, v: Vector) -> bool {
+        *self == v
     }
     fn overlaps(&self, shape: &impl Shape) -> bool {
         shape.contains(*self)
@@ -297,7 +309,7 @@ impl Shape for Vector {
     fn bounding_box(&self) -> Rectangle {
         Rectangle::new(*self, Vector::ONE)
     }
-    fn translate(&self, v: impl Into<Vector>) -> Vector {
-        *self + v.into()
+    fn translate(&self, v: Vector) -> Vector {
+        *self + v
     }
 }